[system admin - departments] fixup & improvements
    - improved the UX for 'member item'
    - added paginator for the member list
    - fixed sorting members by 'role'
    - fixed member name link
    - added confirm dialog for 'delete member'
    - fixed displaying error msgs for adding multiple members
    - removed 'store the sort way in localStorage'
    - removed 'v2' from code & file names
    - cleaned up files. removed files for the previous version, and etc.
    - fixed 'delete repo'
    - improved the scroll of the main content area (make the department
      name bar & the tabs bar fixed)
    - fixup for mobile
